Remove blot, parent, and leaf interfaces
