+++ src/blot/abstract/container.ts
@@ -1,4 +1,3 @@
-import { Blot } from './blot';
 import ParentBlot from './parent';
 import BlockBlot from '../block';
 import * as Registry from '../../registry';
