+++ src/blot/abstract/formattable.ts
@@ -1,0 +1,4 @@
+export interface Formattable {
+  format(name: string, value: any): void;
+  formats(): { [index: string]: any };
+}
