+++ src/registry.ts
@@ -1,5 +1,6 @@
 import Attributor from './attributor/attributor';
-import { Blot, Formattable } from './blot/abstract/blot';
+import { Formattable } from './blot/abstract/formattable';
+import Blot from "./blot/abstract/blot";
 
 export interface BlotConstructor {
   blotName: string;
